Add unit tests covering batch-sinks mode

+++ development/stream_interface/multiplexer_inference_pipeline.py
@@ -1,23 +1,19 @@
 import argparse
+import json
 import os
-import time
-from datetime import datetime
+from dataclasses import asdict
+from datetime import date, datetime
+from enum import Enum
 from functools import partial
 from threading import Thread
-from typing import List
+from typing import Any
 
 import cv2
 import numpy as np
-import supervision as sv
 
 from inference import InferencePipeline
-from inference.core.interfaces.camera.entities import StatusUpdate
-from inference.core.interfaces.camera.utils import multiplex_videos
-from inference.core.interfaces.camera.video_source import VideoSource
-from inference.core.interfaces.stream.inference_pipeline import SinkMode
 from inference.core.interfaces.stream.sinks import render_boxes
-from inference.core.models.utils.batching import create_batches
-from inference.core.utils.preprocess import letterbox_image
+from inference.core.interfaces.stream.watchdog import BasePipelineWatchDog
 
 STREAM_SERVER_URL = os.getenv("STREAM_SERVER", "rtsp://localhost:8554")
 
@@ -30,19 +26,45 @@ def main(n: int) -> None:
     stream_uris = []
     for i in range(n):
         stream_uris.append(f"{STREAM_SERVER_URL}/live{i}.stream")
+    watchdog = BasePipelineWatchDog()
     pipeline = InferencePipeline.init(
         video_reference=stream_uris,
         model_id="yolov8n-640",
         on_prediction=partial(render_boxes, display_statistics=True),
+        watchdog=watchdog,
     )
-    try:
-        pipeline.start()
-    except KeyboardInterrupt:
-        print("Terminating")
-        pipeline.terminate()
-    finally:
-        pipeline.join()
+    control_thread = Thread(target=command_thread, args=(pipeline, watchdog, ))
+    control_thread.start()
+    pipeline.start()
+    pipeline.join()
     cv2.destroyAllWindows()
+
+
+def command_thread(pipeline: InferencePipeline, watchdog: BasePipelineWatchDog) -> None:
+    global STOP
+    while not STOP:
+        try:
+            payload = input()
+            if payload == "s":
+                pipeline.terminate()
+                STOP = True
+            elif payload == "p":
+                pipeline.mute_stream()
+            elif payload == "r":
+                pipeline.resume_stream()
+            elif payload == "i":
+                print(json.dumps(asdict(watchdog.get_report()), default=serialise_to_json,))
+        except Exception as e:
+            print(e)
+    print("END CMD THREAD")
+
+
+def serialise_to_json(obj: Any) -> Any:
+    if isinstance(obj, (datetime, date)):
+        return obj.isoformat()
+    if issubclass(type(obj), Enum):
+        return obj.value
+    raise TypeError(f"Type {type(obj)} not serializable")
 
 
 if __name__ == "__main__":
